requirements.txt
google-genai>=1.0.0
fastapi>=0.110.0
uvicorn>=0.30.0
python-dotenv>=0.21.0
